Added node pattern, checking before issuing a sync across compound node
Previously, a sync token was issued whenever a compound node was called with unique parameters for each constituent node. This created issues with functions that don’t accept sync but might want to be set independently, like setMotorCurrent(). Now, a syncToken type is only generated if the target method/function is able to accept them.